Pulled symbols out into its own module
Pulled symbols out into its own module
