+++ Makefile
@@ -10,7 +10,7 @@ vpath %.h include
 CC = clang
 CPPFLAGS = -std=c99 -Wall -g -I include
 objects = ptscheme.o memmanager.o lispbool.o lispchar.o lispint.o \
-          lispstr.o lisperr.o lisppair.o hashtable.o
+          lispstr.o lisperr.o lisppair.o hashtable.o symbols.o
 testedlibs = hashtable.o
 tests = hashtabletest.o
 
@@ -40,7 +40,11 @@ lisppair.o: ptscheme.h memmanager.h lisppair.h
 
 hashtable.o:
 
+symbols.o: ptscheme.h memmanager.h symbols.h hashtable.h
+
 hashtabletest.o: hashtable.h
+
+
 	clang $(CPPFLAGS) -c -o hashtabletest.o tests/hashtabletest.c
 
 # test - Implement test target
