+++ include/symbols.h
@@ -1,0 +1,9 @@
+/* symbols.h */
+
+#ifndef FILE_SYMBOLS_SEEN
+#define FILE_SYMBOLS_SEEN
+
+extern object *make_symbol(char *value);
+extern char is_symbol(object *obj);
+
+#endif // FILE_SYMBOLS_SEEN
